refactor: changing how the layouts are rendered and removed dashboard route
