+++ web-client/src/types/module.d.ts
@@ -1,3 +1,10 @@
+export interface MenuItem {
+  id: string;
+  icon?: React.FunctionComponent<{}> | React.ComponentClass<{}, any>;
+  title: string;
+  children?: Array<MenuItem>;
+  location?: { path: string };
+}
 export interface Module {
   path: string;
   protected?: boolean;
